+++ app/Http/Controllers/Admin/Product/ProductActivationKey/StoreController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Admin\Product\ProductActivationKey;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Admin\Product\ProductActivationKey\StoreRequest;
+use App\Http\Requests\Admin\ActivationKey\StoreRequest;
 use App\Models\Product;
 
 class StoreController extends Controller
