+++ routes/admin/categories.php
@@ -2,11 +2,13 @@
 
 use Illuminate\Support\Facades\Route;
 
-use App\Http\Controllers\Admin\Category\IndexController;
-use App\Http\Controllers\Admin\Category\StoreController;
-use App\Http\Controllers\Admin\Category\ShowController;
-use App\Http\Controllers\Admin\Category\UpdateController;
-use App\Http\Controllers\Admin\Category\DestroyController;
+use App\Http\Controllers\Admin\Category\{
+  IndexController,
+  StoreController,
+  ShowController,
+  UpdateController,
+  DestroyController
+};
 
 Route::group(['prefix' => 'categories'], function () {
     Route::get('/', IndexController::class);
