+++ routes/admin/products.php
@@ -1,25 +1,32 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Admin\Product\IndexController;
-use App\Http\Controllers\Admin\Product\StoreController;
-use App\Http\Controllers\Admin\Product\ShowController;
-use App\Http\Controllers\Admin\Product\UpdateController;
-use App\Http\Controllers\Admin\Product\DestroyController;
+
+//  Product
+use App\Http\Controllers\Admin\Product\{
+  IndexController,
+  StoreController,
+  ShowController,
+  UpdateController,
+  DestroyController
+};
 
 // ProductActivationKey
-use App\Http\Controllers\Admin\Product\ProductActivationKey\StoreController as ProductActivationKeyStoreController;
-use App\Http\Controllers\Admin\Product\ProductActivationKey\DestroyController as ProductActivationKeyDestroyController;
-use App\Http\Controllers\Admin\Product\ProductActivationKey\IndexController as ProductActivationKeyIndexController;
+use App\Http\Controllers\Admin\Product\ProductActivationKey\{
+  StoreController as ProductActivationKeyIndexController,
+  StoreController as ProductActivationKeyStoreController,
+  DestroyController as ProductActivationKeyDestroyController
+};
 
 
 Route::group(['prefix' => 'products'], function () {
+    //  Product
     Route::get('/', IndexController::class);
     Route::post('/', StoreController::class);
     Route::get('/{product}', ShowController::class);
     Route::patch('/{product}', UpdateController::class);
     Route::delete('/{product}', DestroyController::class);
-
+    //  ProductActivationKey
     Route::prefix('{product}/activation-keys')->group(function(){
         Route::get('/',  ProductActivationKeyIndexController::class);
         Route::post('/', ProductActivationKeyStoreController::class);
