+++ routes/admin/users.php
@@ -1,7 +1,8 @@
 <?php
+
 use Illuminate\Support\Facades\Route;
 
-// User
+//  User
 use App\Http\Controllers\Admin\User\{
     IndexController,
     StoreController,
@@ -10,37 +11,34 @@ use App\Http\Controllers\Admin\User\{
     DestroyController,
 };
 
-// UserOrder
+//  UserOrder
 use App\Http\Controllers\Admin\User\UserOrder\{
     IndexController as UserOrderIndexController,
     StoreController as UserOrderStoreController,
     DestroyController as UserOrderDestroyController,
 };
 
-// UserCart
+//  UserCart
 use App\Http\Controllers\Admin\User\UserCart\{
     ShowController as UserCartShowController,
 };
 
 
-
-
 Route::group(['prefix' => 'users'], function () {
+    //  User
     Route::get('/', IndexController::class);
     Route::post('/', StoreController::class);
     Route::get('/{user}', ShowController::class);
     Route::patch('/{user}', UpdateController::class);
     Route::delete('/{user}', DestroyController::class);
-
-    // UserOrder
+    //  UserOrder
     Route::prefix('{user}/orders')->group(function () {
         Route::get('/', UserOrderIndexController::class);
         Route::post('/', UserOrderStoreController::class);
         Route::delete('/{order}', UserOrderDestroyController::class);
     });
-
-    // UserCart
-    Route::prefix('{user}/cart')->group(function(){
+    //  UserCart
+    Route::prefix('{user}/cart')->group(function () {
         Route::get('/', UserCartShowController::class);
     });
 });
